Make integration test subscription behaviour consistent with JS tests
Make sure that we perform the subscription synchronously so as not to
miss events. This is a nuance that I missed when looking at Cursor's
generated code in fa255c1.

The way in which I've chosen to do this is not particularly elegant (and
to be honest Swift doesn't make it very easy to do so) but I think that
it's the closest to the JS code, which hopefully will mean Cursor will
be able to easily apply this pattern when translating further JS tests.